1) Some fixes

+++ TravelWithPleasureHub/src/main/java/com/kh021j/travelwithpleasurehub/repository/MeetingRepository.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 @Repository
-public interface MeetingRepository extends JpaRepository<Meeting, Long> {
+public interface MeetingRepository extends JpaRepository<Meeting, Integer> {
 
     Meeting findById(long id);
